+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
     - Issue 211 - Query track ingest table for granules with "to_ingest" status
+    - Issue 212 - Update track ingest table with granule status
 ### Changed
 ### Deprecated
 ### Removed